Added error handling to kill, removed old code

+++ src/CS311_A01_2024394_built_in_funcs.c
@@ -50,6 +50,12 @@ void kill2(char **args) {
     }
     
     pid_t target_pid = atoi(args[1]); 
+
+    if (target_pid <= 0) 
+    {
+        fprintf(stderr, "mysh: invalid PID\n");
+        return;
+    }
     
     if (kill(target_pid, SIGTERM) == 0) 
     {
